requirements.txt
Django==5.1.6          # Versión base de Django (ajusta según tu entorno)
Pillow==10.3.0         # Para manejo de imágenes (Image.open, ImageOps)
torch==2.5.1          # PyTorch (modelo de red neuronal)
torchvision==0.20.1    # Transformaciones de imágenes (transforms.Compose)
gunicorn==23.0.0       # Servidor web para produccióngs
numpy==1.26.4          # Operaciones matemáticas
opencv-python==4.11.0 # OpenCV (cv2)
